User functionality aggregated into struct, moved to new base contract (#52)

+++ test/TestQINTokenSale.sol
@@ -15,7 +15,7 @@ contract TestQINTokenSale {
         uint startTime = now + 100;
         uint endTime = now + 200;
         address wallet = 0x1234;
-        uint restrictedDays = 3;
+        uint8 restrictedDays = 3;
         QINToken qin = new QINToken(true);
         QINTokenSale ts = new QINTokenSale(
             qin,
@@ -35,7 +35,7 @@ contract TestQINTokenSale {
         uint startTime = now + 100;
         uint endTime = now + 200;
         address wallet = 0x1234;
-        uint restrictedDays = 3;
+        uint8 restrictedDays = 3;
         QINToken qin = new QINToken(true);
         qin.startTokenSale(
             startTime,
